Fix SSL Secret profile issues
Problem: Some logging around using Secret SSL profiles was confusing or vague. Also, the partition was not included if only the profile name was provided (which is the case with Secrets), resulting in a bunch of unnecessary updates to the virtual server. Finally, configmap profiles were not being cleaned up properly when removed.

Solution: Update the logging to be clearer about what is happening, and removed a log that was misleading. Ensured that if a secret profile name was used, then the partition is added to prevent constant updates to the virtual server. Also ensured that profiles are cleaned up in the configmap case.