Unit test shouldn't look for python driver
Problem: If running locally without the bigipconfigdriver.py installed, one of the unit tests would fail when looking for the existence to this file. This test shouldn't care whether the file exists, as that is not what is being tested.

Solution: Remove the check for the existence of bigipconfigdriver.py, as it is not necessary to the test.