Fixes bug causing periodic failure of the 'Mock driver subprocess' unit test
Problem:
The 'Mock driver subprocess' unit test in main_test.go periodically
fails due to the AfterEach() section attempting to kill the
bigipconfigdriver.py process before both python and this subprocess had
completed starting up. This would cause an unexpected value to return
from the subprocess.

Solution:
The test now waits for the python subprocess to create a temporary file
and for this file to be written to by the subprocess before attemping
to kill the subprocess.

affects-branches: master