Fix node label selector in cluster mode
Problem: The node label selector was ignored in cluster mode due to the endpoints being acquired without caring which node they came from.

Solution: When getting the endpoints for cluster mode, the controller will now check which node they came from. This will ensure that if a node is not being watched, the controller will not use endpoints on that node.