Fixed repeating INFO logs when a pool only config map or ingress is deployed.
Problem:
When a pool only config map or ingress was deployed, the INFO log message
stating "No virtual IP was specified for the virtual server" would get
logged multiple times.

Solution:
This log message is now located in the validResources.go file within the
checkValidIngress and checkValidConfigMap methods rather than in the
resourceConfig.go file. There is now an additional check to guarantee that
the log message is only logged the first time a config is seen.

affects-branch: master