Whitelist Route default SSL profiles
Problem: When cleaning up profiles that no longer exist in a resource, we maintain a whitelist of profiles that don't come from any resource, so we know not to accidentally delete them. The client and server ssl profiles for Routes that are provided via CLI were not considered in this list.

Solution: Whitelist these two profiles to prevent them from being unnecessarily deleted. Also fixed an issue with the configuration of the server ssl profile name.