Updated the k8s controller to match changes in CCCL.
Problem:
CCCL was placing keys and certificates in the Common
partition, but it has been updated to place them in
the managed partition. The k8s controller was also
coded to automatically place these in the Common
partition.

Solution:
The function that sets the path for the keys and
certificates now takes in another arguement for the
name of the managed partition.

affects-branches: master