Delete rules that no longer exist in Ingress
Problem: If a rule was deleted in an Ingress, but the pool still remained (pool was referenced multiple times in Ingress or multi-service gets reduced to single-service), the rule would still be configured.

Solution: Track Ingress host/path rules as an object dependency. If an Ingress gets updated and that host/path rule no longer exists, then the rule is deleted from the config.