Fix multi-service to single-service transition
Problem: When updating a multi-service Ingress to a single-service Ingress, there were a few issues. First, the controller would still have the previous multi-service ingress stored, and so when seeing the single-service, would deny the IP address to be shared, and reject the Ingress. Also, if changing from a multi-service to single-service, the default pool was not updated.

Solution: Change the single-service check to verify whether or not the previous config came from an ingress with the same name, rather than of the same type. This allows different type ingresses to be updated properly. If going from a multi-service Ingress to single-service, set the virtual's default pool.