Ingore versions of pip packages already on rhel host while building
Problem:
If the RHEL host has pip packages that are part of the distribution,
and they conflict with the versions required by the pip install
commands in the Dockerfile, the install fails.

Solution:
Ignore any installed versions of pip packages on the RHEL host
and install the ones specified in the requirements file.