Allow files for BIG-IP parameters
Problem: Using the CLI to provide BIG-IP arguments is a security vulnerability, as they are displayed when running 'ps'.

Solution: Allow for users to create and mount files via a Secret, and provide the mount path to the controller. The controller
will look in this path for username, password, and url files. If they are not found, it falls back to using the CLI.