Don't remove all resources when deleting service
Problem: A bug existed for both Routes and Ingresses where all BIG-IP resources would temporarily be deleted
when a single service was deleted. This was due to some bad logic that would deactivate an entire config if a service for
that config wasn't found.

Solution: The logic now considers if there are multiple services for a config, and some are still active, then don't
deactivate the config if a single service goes away. Only when all services are gone will the config be removed.