Address multiple Services tagged with the same labels
Problem: When service discovery labels are tagged to multiple
Services, all of the endpoints are added to AS3 template. This can
potentially cause issues in traffic distribution.

Solution: Detect services that use the same tags and list them to the user and
don't process any endpoint until there is a one to one mapping between a
Service and Pool.

Affected-branches: feature.user-defined-as3, master