Option to disable AS3 Validation
Problem: We are customizing and hard linking AS3 JSON Schema with the
controller. If there are bugs in the JSON schema or the validator we use,
there is no way to address this situation without shipping a new release.

Solution: Make AS3 Schema validation configurable so that the user can
disable validation.

affected-branches: feature.user-defined-as3