performance improvement: use informer for svc
to get svc using informer way, besides endpointInf.
Because this kind of list was called many times, if using
the old way, it takes more than 1 min to deploy the
configurations onto the bigip when the resource number
gets > 100. Using the new way minimize the time to about
10 seconds.

I used 'go mod vendor' to make vendored copy of dependencies,
otherwise, only service.go and expansion_generated.go are needed.
If we don't need vendor anymore, we can remove that.